# public/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Piyachok API",
    "description": "Документация backend-платформы проекта Piyachok."
  },
  "host": "localhost:3000",
  "basePath": "/api",
...
  "schemes": ["http"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Введіть токен у форматі: Bearer <AccessToken>"
    }
  },
  "paths": {
...
    "/user/profile": {
      "patch": {
        "tags": ["Users"],
        "summary": "Оновлення даних профілю поточного користувача",
        "description": "Дозволяє авторизованому користувачеві змінити свої персональні дані: ім'я, email, телефон чи аватарку. При зміні email перевіряється його унікальність у системі.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": { "type": "string", "example": "Дмитро Коваленко" },
                "email": { "type": "string", "format": "email", "example": "new-email@example.com" },
                "avatarUrl": { "type": "string", "example": "https://s3.amazonaws.com/new-avatar.jpg" },
                "phoneNumber": { "type": "string", "example": "+380679998877" }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Профіль успішно оновлено",
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string", "example": "Профіль успішно оновлено!" },
                "user": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "example": "user-789-uuid" },
                    "name": { "type": "string", "example": "Дмитро Коваленко" },
                    "email": { "type": "string", "example": "new-email@example.com" },
                    "role": { "type": "string", "example": "MANAGER" },
                    "avatarUrl": { "type": "string", "example": "https://s3.amazonaws.com/new-avatar.jpg" },
                    "phoneNumber": { "type": "string", "example": "+380679998877" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Email вже зайнятий чи ні даних для оновлення",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "401": {
            "description": "Неавторизований доступ",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "500": {
            "description": "Внутрішня помилка сервера",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          }
        }
      }
    },
...
    "/venues/{id}/analytics": {
      "get": {
        "tags": ["Venues"],
        "summary": "Отримання статистики переглядів закладу",
        "description": "Повертає агреговану статистику переглядів картки закладу щодня. Доступно лише власнику закладу (менеджеру) або супер-адміністратору.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID закладу, для якого запитується статистика"
          }
        ],
        "responses": {
          "200": {
            "description": "Статистика переглядів успішно отримана",
            "schema": {
              "type": "object",
              "properties": {
                "stats": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "date": { "type": "string", "example": "2026-07-15", "description": "Дата у форматі YYYY-MM-DD" },
                      "count": { "type": "integer", "example": 42, "description": "Кількість унікальних переглядів за цей день" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Неавторизований доступ (відсутня токен)",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "403": {
            "description": "Доступ заборонено (користувач не є власником закладу та не SUPERADMIN)",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "404": {
            "description": "Заклад із зазначеним ID не знайдено",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "500": {
            "description": "Внутрішня помилка сервера",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          }
        }
      }
    },
    "/venues/{id}/news/{newsId}": {
      "patch": {
        "tags": ["Venues"],
        "summary": "Оновлення новини закладу",
        "description": "Дозволяє власнику закладу або супер-адміністратору оновити заголовок, контент, зображення або категорію конкретної новини.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID закладу"
          },
          {
            "name": "newsId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID новини, яку потрібно оновити"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": { "type": "string", "example": "Оновлена назва акції" },
                "content": { "type": "string", "example": "Нові деталі про знижки..." },
                "image": { "type": "string", "x-nullable": true, "example": "https://s3.amazonaws.com/news/123.jpg" },
                "category": { "type": "string", "example": "PROMOTION", "description": "Категорія новини з Enum у Prisma" }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Новину успішно оновлено",
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string", "example": "Новину успішно оновлено!" },
                "news": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "example": "news-555-uuid" },
                    "venueId": { "type": "string", "example": "venue-123-uuid" },
                    "title": { "type": "string", "example": "Оновлена назва акції" },
                    "content": { "type": "string", "example": "Нові деталі про знижки..." },
                    "image": { "type": "string", "example": "https://s3.amazonaws.com/news/123.jpg" },
                    "category": { "type": "string", "example": "PROMOTION" },
                    "createdAt": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Неавторизований доступ або користувач має роль USER",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "403": {
            "description": "Доступ заборонено (користувач не є власником закладу та не SUPERADMIN)",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "404": {
            "description": "Заклад або новину не знайдено (або новина не належить цьому закладу)",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "500": {
            "description": "Внутрішня помилка сервера",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          }
        }
      },
      "delete": {
        "tags": ["Venues"],
        "summary": "Видалення новини закладу",
        "description": "Видаляє новину закладу за її ID. Доступно тільки власнику закладу або супер-адміністратору.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID закладу"
          },
          {
            "name": "newsId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID новини, яку потрібно видалити"
          }
        ],
        "responses": {
          "200": {
            "description": "Новину успішно видалено",
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string", "example": "Новину успішно видалено!" }
              }
            }
          },
          "401": {
            "description": "Неавторизований доступ",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "403": {
            "description": "Доступ заборонено (користувач не є власником закладу та не SUPERADMIN)",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "404": {
            "description": "Заклад або новину не знайдено",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "500": {
            "description": "Внутрішня помилка сервера",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          }
        }
      }
    },
...
    "/venues/{id}/reviews": {
      "post": {
        "tags": ["Reviews"],
        "summary": "Додавання відгуку або скарги на заклад",
        "description": "Дозволяє авторизованому користувачу залишити відгук з оцінкою від 1 до 5 або надіслати скаргу модераторам. Метод автоматично перераховує середній рейтинг закладу в базі даних всередині ACID-транзакції.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID закладу, для якого створюється відгук"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["rating"],
              "properties": {
                "rating": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 5,
                  "example": 5,
                  "description": "Оцінка закладу від 1 до 5 зірок"
                },
                "text": {
                  "type": "string",
                  "example": "Чудове місце, великий вибір крафту та приємна атмосфера!",
                  "description": "Текст відгуку (необов'язковий)"
                },
                "isComplaint": {
                  "type": "boolean",
                  "example": false,
                  "description": "Прапор, чи є цей відгук скаргою (наприклад, на обман, невідповідність опису)"
                },
                "complaintReason": {
                  "type": "string",
                  "x-nullable": true,
                  "example": "Заклад зачинено назавжди за цією адресою",
                  "description": "Причина скарги (обов'язкова, якщо isComplaint: true)"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Відгук успішно створено, а рейтинг закладу автоматично перераховано",
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string", "example": "Відгук успішно додано, рейтинг закладу оновлено!" },
                "review": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "example": "review-999-uuid" },
                    "rating": { "type": "integer", "example": 5 },
                    "text": { "type": "string", "example": "Чудове місце..." },
                    "userId": { "type": "string", "example": "user-789-uuid" },
                    "venueId": { "type": "string", "example": "venue-123-uuid" },
                    "isComplaint": { "type": "boolean", "example": false },
                    "complaintReason": { "type": "string", "x-nullable": true },
                    "createdAt": { "type": "string", "format": "date-time" },
                    "user": {
                      "type": "object",
                      "properties": {
                        "name": { "type": "string", "example": "Олександр" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Помилка валідації (оцінка відсутня або виходить за межі 1-5)",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "401": {
            "description": "Неавторизований доступ (користувач не увійшов у систему або заблокований)",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "404": {
            "description": "Заклад з вказаним ID не знайдено",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "500": {
            "description": "Внутрішня помилка сервера при виконанні транзакції",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "message": { "type": "string", "example": "Внутрішня помилка сервера / Доступ заборонено" }
      }
    },
...
  }
}